requirements.txt
# 核心依赖
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 数据库
sqlalchemy>=2.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0
impyla>=0.16.1

# 时间序列模型
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0
prophet>=1.1.0
statsmodels>=0.14.0

# AutoML & 超参数优化
optuna>=3.3.0
# autogluon.tabular>=0.8.0  # 可选：需要时解除注释

# 模型管理
mlflow>=2.8.0

# API 服务
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 任务调度
apscheduler>=3.10.0

# 数据处理
pyarrow>=13.0.0
openpyxl>=3.1.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# 监控
prometheus-client>=0.18.0

# 工具
tqdm>=4.66.0
joblib>=1.3.0
python-dateutil>=2.8.0

# 日志
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
